+++ CHANGELOG.md
@@ -10,8 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 Added:
 
 * http: json key/val equality check, using JMESPath
-* metrics filesystem: collect metrics for mounted filesystems
 * time: NTP check / metrics
+
+Changed:
+
+* filesystem: collect metrics for mounted filesystems
+* http: CURL-like metrics
 
 ## [v0.1.4] - Dec 13, 2019