+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+No changes so far.
+
+## [v0.2.0] - Jan 19, 2020
+
 Added:
 
 * http: json key/val equality check, using JMESPath
@@ -70,6 +74,7 @@ Added:
 * http check
 
 [Unreleased]: https://github.com/julian7/sensu-base-checks
+[v0.1.4]: https://github.com/julian7/sensu-base-checks/releases/tag/v0.2.0
 [v0.1.4]: https://github.com/julian7/sensu-base-checks/releases/tag/v0.1.4
 [v0.1.3]: https://github.com/julian7/sensu-base-checks/releases/tag/v0.1.3
 [v0.1.2]: https://github.com/julian7/sensu-base-checks/releases/tag/v0.1.2